+++ config/landov3.js
@@ -272,6 +272,12 @@ export default function({
       team: [],
     },
     vite: {
+      css: {
+        preprocessorOptions: {
+          sass: {api: 'modern-compiler'},
+          scss: {api: 'modern-compiler'},
+        },
+      },
       optimizeDeps: {exclude: []},
       plugins: [],
       resolve: {alias: []},