+++ config/landov4.js
@@ -253,6 +253,12 @@ export default function({base, landoPlugin, themeConfig, version}) {
       team: [],
     },
     vite: {
+      css: {
+        preprocessorOptions: {
+          sass: {api: 'modern-compiler'},
+          scss: {api: 'modern-compiler'},
+        },
+      },
       optimizeDeps: {exclude: []},
       plugins: [],
       resolve: {alias: []},